Move formatter to seperate file. Rework GetMeasurement
