+++ src/openrct2/libopenrct2.vcxproj
@@ -219,6 +219,7 @@
     <ClInclude Include="localisation\Currency.h" />
     <ClInclude Include="localisation\Date.h" />
     <ClInclude Include="localisation\FormatCodes.h" />
+    <ClInclude Include="localisation\Formatter.h" />
     <ClInclude Include="localisation\Language.h" />
     <ClInclude Include="localisation\LanguagePack.h" />
     <ClInclude Include="localisation\Localisation.h" />
@@ -546,6 +547,7 @@
     <ClCompile Include="localisation\Convert.cpp" />
     <ClCompile Include="localisation\Currency.cpp" />
     <ClCompile Include="localisation\FormatCodes.cpp" />
+    <ClCompile Include="localisation\Formatter.cpp" />
     <ClCompile Include="localisation\Language.cpp" />
     <ClCompile Include="localisation\LanguagePack.cpp" />
     <ClCompile Include="localisation\Localisation.cpp" />
