+++ src/openrct2/localisation/Localisation.h
@@ -17,8 +17,6 @@
 #include "Language.h"
 #include "StringIds.h"
 
-#include <array>
-#include <cstring>
 #include <string>
 
 bool utf8_is_format_code(char32_t codepoint);
@@ -67,7 +65,6 @@ extern const char real_name_initials[16];
 extern const char* real_names[1024];
 
 extern thread_local char gCommonStringFormatBuffer[512];
-extern thread_local uint8_t gCommonFormatArgs[80];
 extern bool gDebugStringFormatting;
 
 extern const rct_string_id SpeedNames[5];
@@ -79,111 +76,5 @@ extern const rct_string_id PeepThoughts[174];
 extern const rct_string_id DateDayNames[31];
 extern const rct_string_id DateGameMonthNames[MONTH_COUNT];
 extern const rct_string_id DateGameShortMonthNames[MONTH_COUNT];
-
-class Formatter
-{
-    std::array<uint8_t, 80> Buffer{};
-    uint8_t* StartBuf{};
-    uint8_t* CurrentBuf{};
-
-public:
-    explicit Formatter(uint8_t* buf)
-        : StartBuf(buf)
-        , CurrentBuf(buf)
-    {
-    }
-
-    Formatter()
-        : Buffer{}
-        , StartBuf(Buffer.data())
-        , CurrentBuf(StartBuf)
-    {
-    }
-
-    Formatter(const Formatter& other)
-    {
-        *this = other;
-    }
-
-    Formatter& operator=(const Formatter& other)
-    {
-        // If using global or not
-        if (other.StartBuf == other.Buffer.data())
-        {
-            std::copy(std::begin(other.Buffer), std::end(other.Buffer), std::begin(Buffer));
-            StartBuf = Buffer.data();
-        }
-        else
-        {
-            StartBuf = other.StartBuf;
-        }
-        CurrentBuf = StartBuf + other.NumBytes();
-        return *this;
-    }
-
-    static Formatter Common()
-    {
-        return Formatter{ gCommonFormatArgs };
-    }
-
-    auto Buf()
-    {
-        return CurrentBuf;
-    }
-
-    auto Data() const
-    {
-        return StartBuf;
-    }
-
-    void Increment(size_t count)
-    {
-        CurrentBuf += count;
-    }
-
-    void Rewind()
-    {
-        CurrentBuf -= NumBytes();
-    }
-
-    std::size_t NumBytes() const
-    {
-        return CurrentBuf - StartBuf;
-    }
-
-    template<typename TSpecified, typename TDeduced> Formatter& Add(TDeduced value)
-    {
-        static_assert(sizeof(TSpecified) <= sizeof(uintptr_t), "Type too large");
-        static_assert(sizeof(TDeduced) <= sizeof(uintptr_t), "Type too large");
-
-        // clang-format off
-        static_assert(
-            std::is_same_v<TSpecified, char*> ||
-            std::is_same_v<TSpecified, const char*> ||
-            std::is_same_v<TSpecified, int16_t> ||
-            std::is_same_v<TSpecified, int32_t> ||
-            std::is_same_v<TSpecified, money32> ||
-            std::is_same_v<TSpecified, rct_string_id> ||
-            std::is_same_v<TSpecified, uint16_t> ||
-            std::is_same_v<TSpecified, uint32_t> ||
-            std::is_same_v<TSpecified, utf8*> ||
-            std::is_same_v<TSpecified, const utf8*>
-        );
-        // clang-format on
-
-        uintptr_t convertedValue;
-        if constexpr (std::is_integral_v<TSpecified>)
-        {
-            convertedValue = static_cast<uintptr_t>(value);
-        }
-        else
-        {
-            convertedValue = reinterpret_cast<uintptr_t>(value);
-        }
-        std::memcpy(CurrentBuf, &convertedValue, sizeof(TSpecified));
-        Increment(sizeof(TSpecified));
-        return *this;
-    }
-};
 
 #endif
